Add files via upload
The IOT based Garbage monitoring system is a very advanced system which helps to keep the cities clean. This system monitors the garbage bins and conveys the level of garbage collected in the dustbins to Municipal corporations through GPS tracker present inside these system. Here ULTRASONIC sensor is used to track the level of garbage collected inside bins.